// frontend/src/api/generated/models/meuSemestreDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import type { SemestreResumoDto } from './semestreResumoDto';
import type { TurmaResumoMeuSemestreDto } from './turmaResumoMeuSemestreDto';

export interface MeuSemestreDto {
  semestre: SemestreResumoDto;
  turma: TurmaResumoMeuSemestreDto;
  /** Se é o semestre da turma atual do aluno */
  atual: boolean;
  totalMaterias: number;
  aprovadas: number;
  reprovadas: number;
  cursando: number;
}